pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depth_estimation"
version = "0.1.1"
authors = [
    { name = "shriarul5273" },
]
description = "A Transformers-style Python library for monocular depth estimation — inference, evaluation, and fine-tuning"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
keywords = ["depth-estimation", "monocular-depth", "computer-vision", "deep-learning", "pytorch", "fine-tuning", "training"]
dependencies = [
    "torch==2.9.0",
    "torchvision>=0.15",
    "Pillow>=9.0",
    "numpy>=1.24",
    "matplotlib>=3.6",
    "opencv-python>=4.8",
    "huggingface-hub>=0.16",
    "transformers",
    "diffusers",
    "accelerate",
    "timm>=0.9.1",
    "einops>=0.6",
    "addict",
    "h5py",
    "tqdm"
]

[project.urls]
Homepage = "https://github.com/shriarul5273/depth_estimation"
Repository = "https://github.com/shriarul5273/depth_estimation"
Issues = "https://github.com/shriarul5273/depth_estimation/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
depth-estimate = "depth_estimation.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
